+++ spec/unit/recipes/default_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-describe 'verifi_skeleton::default' do
+describe '_logstash_server3_rm::default' do
   let(:chef_run) { ChefSpec::Runner.new.converge(described_recipe) }
 
-  it 'example skeleton test' do
-    expect(chef_run).to include_recipe('verifi_skeleton::default')
+  it 'example logstash_server3_rm test' do
+    expect(chef_run).to include_recipe('_logstash_server3_rm::default')
   end
 end